+++ app/Http/Controllers/MataPelajaranController.php
@@ -16,16 +16,20 @@ class MataPelajaranController extends Controller
      */
     public function index()
     {
-        if (Auth::guard('guru')->user()->level == 'tata usaha' || Auth::guard('guru')->user()->level == 'wali kelas') {
-            $title = "Mata Pelajaran";
-            $mapel = MataPelajaran::all();
+        if (Auth::guard('guru')->check()) {
+            if (Auth::guard('guru')->user()->level == 'tata usaha' || Auth::guard('guru')->user()->level == 'wali kelas') {
+                $title = "Mata Pelajaran";
+                $mapel = MataPelajaran::all();
 
-            confirmDelete();
+                confirmDelete();
 
-            return view('dashboard.MataPelajaran.DataMataPelajaran', [
-                'title' => $title,
-                'mapel' => $mapel,
-            ]);
+                return view('dashboard.MataPelajaran.DataMataPelajaran', [
+                    'title' => $title,
+                    'mapel' => $mapel,
+                ]);
+            } else {
+                return back();
+            }
         } else {
             return back();
         }
@@ -36,12 +40,16 @@ class MataPelajaranController extends Controller
      */
     public function create()
     {
-        if (Auth::guard('guru')->user()->level == 'tata usaha') {
-            $title = "Tambah Mata Pelajaran";
+        if (Auth::guard('guru')->check()) {
+            if (Auth::guard('guru')->user()->level == 'tata usaha') {
+                $title = "Tambah Mata Pelajaran";
 
-            return view('dashboard.MataPelajaran.TambahDataMataPelajaran', [
-                'title' => $title,
-            ]);
+                return view('dashboard.MataPelajaran.TambahDataMataPelajaran', [
+                    'title' => $title,
+                ]);
+            } else {
+                return back();
+            }
         } else {
             return back();
         }
@@ -52,33 +60,37 @@ class MataPelajaranController extends Controller
      */
     public function store(Request $request)
     {
-        if (Auth::guard('guru')->user()->level == 'tata usaha') {
+        if (Auth::guard('guru')->check()) {
+            if (Auth::guard('guru')->user()->level == 'tata usaha') {
 
-            $messages = [
-                'required' => ':Attribute harus diisi.',
-                'nama_pelajaran.unique' => 'Mata Pelajaran sudah terdaftar...',
-                'kd_pelajaran.unique' => 'Kode mata pelajaran sudah terdaftar...'
-            ];
+                $messages = [
+                    'required' => ':Attribute harus diisi.',
+                    'nama_pelajaran.unique' => 'Mata Pelajaran sudah terdaftar...',
+                    'kd_pelajaran.unique' => 'Kode mata pelajaran sudah terdaftar...'
+                ];
 
-            $validator = Validator::make($request->all(), [
-                'nama_pelajaran' => 'required|unique:mata_pelajarans,nama_pelajaran',
-                'kd_pelajaran' => 'required|unique:mata_pelajarans,kd_pelajaran',
-            ], $messages);
+                $validator = Validator::make($request->all(), [
+                    'nama_pelajaran' => 'required|unique:mata_pelajarans,nama_pelajaran',
+                    'kd_pelajaran' => 'required|unique:mata_pelajarans,kd_pelajaran',
+                ], $messages);
 
-            if ($validator->fails()) {
-                return redirect()->back()->withErrors($validator)->withInput();
+                if ($validator->fails()) {
+                    return redirect()->back()->withErrors($validator)->withInput();
+                }
+
+                // Sweet alert
+                Alert::success('Berhasil Ditambahkan', 'Mata pelajaran berhasil ditambahkan.');
+
+                // ELOQUENT
+                $mata_pelajaran = new MataPelajaran;
+                $mata_pelajaran->nama_pelajaran = $request->nama_pelajaran;
+                $mata_pelajaran->kd_pelajaran = $request->kd_pelajaran;
+                $mata_pelajaran->save();
+                return redirect()->route('matapelajaran.index');
+
+            } else {
+                return back();
             }
-
-            // Sweet alert
-            Alert::success('Berhasil Ditambahkan', 'Mata pelajaran berhasil ditambahkan.');
-
-            // ELOQUENT
-            $mata_pelajaran = new MataPelajaran;
-            $mata_pelajaran->nama_pelajaran = $request->nama_pelajaran;
-            $mata_pelajaran->kd_pelajaran = $request->kd_pelajaran;
-            $mata_pelajaran->save();
-            return redirect()->route('matapelajaran.index');
-
         } else {
             return back();
         }
@@ -97,14 +109,18 @@ class MataPelajaranController extends Controller
      */
     public function edit(string $id)
     {
-        if (Auth::guard('guru')->user()->level == 'tata usaha') {
+        if (Auth::guard('guru')->check()) {
+            if (Auth::guard('guru')->user()->level == 'tata usaha') {
 
-            $title = "Edit Data Mata Pelajaran";
-            $mapel = MataPelajaran::find($id);
-            return view('dashboard.MataPelajaran.EditDataMataPelajaran', [
-                'title' => $title,
-                'mapel' => $mapel,
-            ]);
+                $title = "Edit Data Mata Pelajaran";
+                $mapel = MataPelajaran::find($id);
+                return view('dashboard.MataPelajaran.EditDataMataPelajaran', [
+                    'title' => $title,
+                    'mapel' => $mapel,
+                ]);
+            } else {
+                return back();
+            }
         } else {
             return back();
         }
@@ -115,32 +131,36 @@ class MataPelajaranController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        if (Auth::guard('guru')->user()->level == 'tata usaha') {
+        if (Auth::guard('guru')->check()) {
+            if (Auth::guard('guru')->user()->level == 'tata usaha') {
 
-            $messages = [
-                'required' => ':Attribute harus diisi.',
-                'email' => 'Isi :attribute dengan format yang benar',
-                'numeric' => 'Isi attribute dengan angka'
-            ];
+                $messages = [
+                    'required' => ':Attribute harus diisi.',
+                    'email' => 'Isi :attribute dengan format yang benar',
+                    'numeric' => 'Isi attribute dengan angka'
+                ];
 
-            $validator = Validator::make($request->all(), [
-                'nama_pelajaran' => 'required',
-                'kd_pelajaran' => 'required',
-            ], $messages);
+                $validator = Validator::make($request->all(), [
+                    'nama_pelajaran' => 'required',
+                    'kd_pelajaran' => 'required',
+                ], $messages);
 
-            if ($validator->fails()) {
-                return redirect()->back()->withErrors($validator)->withInput();
+                if ($validator->fails()) {
+                    return redirect()->back()->withErrors($validator)->withInput();
+                }
+
+                $mata_pelajaran = MataPelajaran::findOrFail($id);
+                $mata_pelajaran->nama_pelajaran = $request->nama_pelajaran;
+                $mata_pelajaran->kd_pelajaran = $request->kd_pelajaran;
+                $mata_pelajaran->save();
+
+                // Sweet alert
+                Alert::success('Perubahan Berhasil', 'Mata pelajaran berhasil diubah.');
+
+                return redirect()->route('matapelajaran.index');
+            } else {
+                return back();
             }
-
-            $mata_pelajaran = MataPelajaran::findOrFail($id);
-            $mata_pelajaran->nama_pelajaran = $request->nama_pelajaran;
-            $mata_pelajaran->kd_pelajaran = $request->kd_pelajaran;
-            $mata_pelajaran->save();
-
-            // Sweet alert
-            Alert::success('Perubahan Berhasil', 'Mata pelajaran berhasil diubah.');
-
-            return redirect()->route('matapelajaran.index');
         } else {
             return back();
         }
@@ -151,15 +171,19 @@ class MataPelajaranController extends Controller
      */
     public function destroy(string $id)
     {
-        if (Auth::guard('guru')->user()->level == 'tata usaha') {
+        if (Auth::guard('guru')->check()) {
+            if (Auth::guard('guru')->user()->level == 'tata usaha') {
 
-            $mapel = MataPelajaran::findOrFail($id);
-            $mapel->delete();
+                $mapel = MataPelajaran::findOrFail($id);
+                $mapel->delete();
 
-            // Sweet alert
-            Alert::success('Berhasil Dihapus', 'Mata pelajaran berhasil dihapus.');
+                // Sweet alert
+                Alert::success('Berhasil Dihapus', 'Mata pelajaran berhasil dihapus.');
 
-            return redirect()->route('matapelajaran.index');
+                return redirect()->route('matapelajaran.index');
+            } else {
+                return back();
+            }
         } else {
             return back();
         }
